extra.css

Add a Jumbotron to showcase key content

Add a Jumbotron to showcase key content

Usage

jumbotron(header, content, button = TRUE, ...)

Arguments

header

a object of class character, indicating header title

content

a object of class character, indicating text content

button

a object of class logical, to include a button or not

...

optional parameters for including a button label if needed

Value

a HTML object to be included in the ui section of a shiny app

See also

Examples

 jumbotron("Hi Shiny!", "text to show", buttonLabel = "Click Me")
#> <div class="jumbotron">
#>   <h1>Hi Shiny!</h1>
#>   <p>text to show</p>
#>   <p>
#>     <a class="btn btn-primary btn-lg button" id="tabBut">Click Me</a>
#>   </p>
#> </div>