Add a Single List Group Component A flexible method for displaying simple list elements as a group.

Add a Single List Group Component A flexible method for displaying simple list elements as a group.

Usage

list_group(list_items)

Arguments

list_items

a vector of content generated from the list_item function

Value

a HTML object to be included in the ui section of a shiny app

See also

Examples

list_group(list_item("Howdy Title", badge_value = 27))
#> <ul class="list-group">
#>   <li class="list-group-item">
#>     <span class="badge">27</span>
#>     Howdy Title
#>   </li>
#> </ul>