extra.css

Add a Single List Item Component A flexible method for displaying simple list elements.

Add a Single List Item Component A flexible method for displaying simple list elements.

Usage

list_item(content, badge = TRUE, badge_value)

Arguments

content

a object of class character

badge

a object of class logical

badge_value

if the badge object is true provide an arguement object of class numeric, named badge_value.

Value

a HTML object to be included in the ui section of a shiny app

See also

Examples

list_item("Howdy Title", badge_value = 27)
#> <li class="list-group-item">
#>   <span class="badge">27</span>
#>   Howdy Title
#> </li>